feat(web): public shell — defer Supabase/auth until post-navigation
Rebase onto develop; removes stale AuthShell.test.tsx from merge.
