+++ apps/web/src/AuthenticatedApp.tsx
@@ -1,20 +1,13 @@
-import { BrowserRouter } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import { AuthProvider } from '@beakerstack/shared/contexts/AuthContext';
 import { ProfileProvider } from '@beakerstack/shared/contexts/ProfileContext';
 import { supabase } from './lib/supabase';
-import App from './App';
 
-interface AuthShellProps {
-  basePath: string;
-}
-
-export function AuthShell({ basePath }: AuthShellProps) {
+export function AuthenticatedApp() {
   return (
     <AuthProvider supabaseClient={supabase}>
       <ProfileProvider supabaseClient={supabase}>
-        <BrowserRouter basename={basePath}>
-          <App />
-        </BrowserRouter>
+        <Outlet />
       </ProfileProvider>
     </AuthProvider>
   );
