+++ apps/web/src/pages/__tests__/HomePage.test.tsx
@@ -200,12 +200,4 @@ describe('HomePage', () => {
     });
   });
 
-  describe('when user is authenticated', () => {
-    it('redirects to /dashboard', async () => {
-      await renderWithAuth(true);
-      await waitFor(() => {
-        expect(screen.getByText('Dashboard')).toBeInTheDocument();
-      });
-    });
-  });
 });
