+++ apps/web/vite.config.ts
@@ -108,9 +108,6 @@ export default defineConfig(({ mode }) => {
           // SSR-only landing component — structural duplicate of LandingPage used by the
           // prerender script only; covered by the build-time prerender smoke check
           'src/components/landing/LandingPageSSR.tsx',
-          // Thin composition wrapper for auth providers — no logic; constituent
-          // providers and routing are tested independently
-          'src/AuthShell.tsx',
           // Display-only dashboard showcase components — no business logic;
           // annotated UI primitives covered visually by preview deployment
           'src/components/dashboard/AnnotatedPrimitive.tsx',
